Further randomize media spawn location
Media now spawns around the screen instead of in the four corners.
A quadrant of the screen is picked at random (ensuring it's not the
same as before) and a random set of coordinates inside the quadrant is
picked. These coordinates are assigned to the image before it appears
on screen.