Fix issue with media showing offscreen on load (#14)
Also makes the media show up in the top left corner of the screen on
page/source load. This is a two-for-one change, as it helps users see
that the app is working by showing them a worm on first load, but also
forces the mediaHeight and mediaWidth variables to have a value during
the first "real" run of the loop.